ml_service/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
xgboost>=2.0.0
scikit-learn>=1.3.0
requests>=2.31.0
joblib>=1.3.0
python-dotenv>=1.0.0

